i2c: Attempt to fix Rusty_CryptoAuthLib
In the PR #148, a quite amount of changes were made. Although each
single line matters, it is not convincing without any unit tests or
working example during review process.

The problem is that Rusty_CryptoAuthLib is confirmed to be broken.
If I'm lucky enough, it won't be a wholesale rewrite or C-binding.